Refactored java packages structure ofr Inventory Device Management Service
