+++ service/device/management/inventory/api/src/main/java/org/eclipse/kapua/service/device/management/inventory/model/packages/DeviceInventoryPackages.java
@@ -10,7 +10,7 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.device.management.inventory.model.bundle.inventory;
+package org.eclipse.kapua.service.device.management.inventory.model.packages;
 
 import org.eclipse.kapua.KapuaSerializable;
 
@@ -22,37 +22,37 @@ import javax.xml.bind.annotation.XmlType;
 import java.util.List;
 
 /**
- * {@link DeviceInventoryBundles} definition.
+ * {@link DeviceInventoryPackages} definition.
  *
  * @since 1.5.0
  */
+@XmlRootElement(name = "deviceInventoryDeploymentPackages")
 @XmlAccessorType(XmlAccessType.PROPERTY)
-@XmlRootElement(name = "deviceInventoryBundles")
-@XmlType(factoryClass = DeviceInventoryBundlesXmlRegistry.class, factoryMethod = "newDeviceInventoryBundles")
-public interface DeviceInventoryBundles extends KapuaSerializable {
+@XmlType(factoryClass = DeviceInventoryPackagesXmlRegistry.class, factoryMethod = "newDeviceInventoryPackages")
+public interface DeviceInventoryPackages extends KapuaSerializable {
 
     /**
-     * Gets the {@link List} of {@link DeviceInventoryBundle}s
+     * Gets the {@link List} of {@link DeviceInventoryPackage}s.
      *
-     * @return The {@link List} of {@link DeviceInventoryBundle}s
+     * @return The {@link List} of {@link DeviceInventoryPackage}s.
      * @since 1.5.0
      */
-    @XmlElement(name = "inventoryBundles")
-    List<DeviceInventoryBundle> getInventoryBundles();
+    @XmlElement(name = "deploymentPackages")
+    List<DeviceInventoryPackage> getPackages();
 
     /**
-     * Adds a {@link DeviceInventoryBundle} to the {@link List}
+     * Adds a {@link DeviceInventoryPackage} to the {@link List}.
      *
-     * @param inventoryBundle The {@link DeviceInventoryBundle} to add.
+     * @param aPackage The {@link DeviceInventoryPackage} to add.
      * @since 1.5.0
      */
-    void addInventoryBundle(DeviceInventoryBundle inventoryBundle);
+    void addPackage(DeviceInventoryPackage aPackage);
 
     /**
-     * Sets the {@link List} of {@link DeviceInventoryBundle}s
+     * Sets the {@link List} of {@link DeviceInventoryPackage}s.
      *
-     * @param inventoryBundles The {@link List} of {@link DeviceInventoryBundle}s
+     * @param inventoryPackages The {@link List} of {@link DeviceInventoryPackage}s.
      * @since 1.5.0
      */
-    void setInventoryBundles(List<DeviceInventoryBundle> inventoryBundles);
+    void setPackages(List<DeviceInventoryPackage> inventoryPackages);
 }
