+++ service/device/management/inventory/internal/src/main/java/org/eclipse/kapua/service/device/management/inventory/internal/message/InventorySystemPackagesResponseMessage.java
@@ -10,15 +10,15 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.device.management.inventory.message.internal;
+package org.eclipse.kapua.service.device.management.inventory.internal.message;
 
-import org.eclipse.kapua.service.device.management.inventory.model.inventory.DeviceInventory;
+import org.eclipse.kapua.service.device.management.inventory.model.system.DeviceInventorySystemPackages;
 import org.eclipse.kapua.service.device.management.message.response.KapuaResponseMessage;
 
 /**
- * {@link DeviceInventory} {@link KapuaResponseMessage} implementation.
+ * {@link DeviceInventorySystemPackages} {@link KapuaResponseMessage} implementation.
  *
  * @since 1.5.0
  */
-public class InventoryListResponseMessage extends InventoryResponseMessage {
+public class InventorySystemPackagesResponseMessage extends InventoryResponseMessage {
 }
