+++ translator/kapua/kura/src/main/java/org/eclipse/kapua/translator/kura/kapua/inventory/TranslatorAppInventoryBundlesKuraKapua.java
@@ -15,8 +15,8 @@ package org.eclipse.kapua.translator.kura.kapua.inventory;
 import org.eclipse.kapua.service.device.call.kura.model.inventory.bundles.KuraInventoryBundles;
 import org.eclipse.kapua.service.device.call.message.kura.app.response.KuraResponseMessage;
 import org.eclipse.kapua.service.device.call.message.kura.app.response.KuraResponsePayload;
-import org.eclipse.kapua.service.device.management.inventory.message.internal.InventoryResponsePayload;
-import org.eclipse.kapua.service.device.management.inventory.message.internal.InventoryBundlesResponseMessage;
+import org.eclipse.kapua.service.device.management.inventory.internal.message.InventoryBundlesResponseMessage;
+import org.eclipse.kapua.service.device.management.inventory.internal.message.InventoryResponsePayload;
 import org.eclipse.kapua.translator.Translator;
 import org.eclipse.kapua.translator.exception.InvalidPayloadException;
